requirements.txt
streamlit>=1.31,<1.40
numpy>=1.24,<2.0
pandas>=1.5,<2.1
requests>=2.31,<3.0
Pillow>=10.0,<11.0
matplotlib>=3.7,<3.9
opencv-python-headless>=4.8,<4.11
PyMuPDF>=1.23,<1.25
pytesseract>=0.3.10,<0.4
lifelines>=0.27,<0.29
